Better input format for shapes (#292)
Refactoring of the input for shapes in the chemiscope input, as well as all related helpers and python utilities. 

Main idea is that now options don't need to be repeated, but can be given globally and then only differences can be specified ad the structure and/or atom level. 

Also added a native Arrow shape, fixed many bugs and improved the documentation. The example contains also a funny (?) easter egg.

---------

Co-authored-by: Rose K. Cersonsky <47536110+rosecers@users.noreply.github.com>
Co-authored-by: Guillaume Fraux <guillaume.fraux@epfl.ch>