Fix symbol and map glitches (#383)
Fixes #380 - or at least, patches up the most glaring problem and adds a
bunch of symbols by monkey-patching plotly. The symbols are glyphs from
fonts, so size depends on which font is picked that seems to be browser
dependent. Also tackles a couple of small glitches when switching back
and forth from 3D/2D